+++ examples/web-components-with-react.jsx
@@ -7,9 +7,9 @@ function HomePage() {
     return (
         <section className="home-page">
             <h1>Hello World</h1>
-            <p>This page is used to show how DataFormsJS Web Components (and Web Components in general) can be used with React and React Router.</p>
+            <p>This page is used to show how DataFormsJS Web Components (and Web Components in general) can be used with React (or Preact) and React Router.</p>
             <p>All Web Component and all page features are expected to work in virtually all browsers when using React.</p>
-            <p>The JSX is shared with both React and Preact demos and currently when using Preact not all pages work.</p>
+            <p>The JSX is shared with both React and Preact demos.</p>
         </section>
     );
 }