+++ test/js/unit-testing.js
@@ -5,7 +5,7 @@
  *
  * Unit Testing runs in the Browser using QUnit.
  * See instructions in [../server.js].
- * 
+ *
  * NOTE - Vue was added several years after this file was creatd
  * and because the behavior of Vue is different from Templating (Handlebars, etc)
  * many tests and checks have to be skipped for Vue. Because many tests
@@ -53,6 +53,8 @@
             assert.deepEqual(app.settings.requestHeadersByHostName, {}, 'Default Settings for app.settings.requestHeadersByHostName: ' + JSON.stringify(app.settings.requestHeadersByHostName));
             assert.equal(app.settings.errors.pageLoading, 'Error loading the current page because the previous page is still loading and is taking a long time. Please refresh the page and try again.', 'Default settings for app.settings.errors.pageLoading: ' + app.settings.errors.pageLoading);
             assert.equal(Object.keys(app.settings).length, 11, 'Number of properties in app.settings');
+            var version = app.version.match(/^5.\d+.\d+$/);
+            assert.ok(version !== null, 'app.version is major version 5, full version: ' + version);
 
             switch (app.viewEngine()) {
                 case 'Handlebars':