+++ CHANGELOG.md
@@ -4,6 +4,14 @@ DataFormsJS uses [Semantic Versioning](https://docs.npmjs.com/about-semantic-ver
 
 Overall the core Framework files and API are expected to remain stable however the version number is expected to increase to much larger numbers in the future due to the changes to smaller scripts and components. This change log includes Framework release history and new website features or major changes.
 
+## Next Release (changes on master branch and not published to npm)
+
+* Web Component `<json-data>`
+  * Added setter properties for `url` and `urlParams`. Previously only getters were defined and this resulted in the Component not working with Preact.
+  * With this release DataFormsJS Web Components now work with Preact in addition to React
+  * Demo:
+    * https://www.dataformsjs.com/examples/web-components-with-preact.htm
+
 ## 5.4.1 (December 9, 2020)
 
 * Web Component `<markdown-content>`